+++ src/auth/oauth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
-import { OAuthProvider, User } from '@prisma/client';
+import { User } from '@prisma/client';
 import { UsersService } from 'src/users/users.service';
 import { OAuthInput } from './interfaces/oauth.interface';
 
@@ -8,8 +8,8 @@ export class OAuthService {
   constructor(private readonly usersService: UsersService) {}
 
   async resolveUser(input: OAuthInput): Promise<User> {
-    const { provider, providerId, email, emailVerified, displayName } = input;
-
+    const { provider, providerId } = input;
+    // 1. 이미 OAath 계정이 연결되어 있는 경우 -> 바로 user 반환
     const account = await this.usersService.findOAuthAccount(
       provider,
       providerId,
@@ -25,49 +25,12 @@ export class OAuthService {
       return user;
     }
 
-    let user: User | null = email
-      ? await this.usersService.findByEmail(email)
-      : null;
+    // 2. 없으면 UserService에게 OAuthInput으로 유저를 찾아/만들어줘 요청
+    const user = await this.usersService.getOrCreateUserFromOAuth(input);
 
-    if (!user) {
-      const username = this.generateUsername(provider, providerId);
-      const resolvedDisplayName = (() => {
-        if (displayName) {
-          return displayName.trim().slice(0, 50);
-        }
-        if (email) {
-          return email.split('@')[0].slice(0, 50);
-        }
-        return `${provider.toString().toLowerCase()}_user`;
-      })();
-      user = await this.usersService.createUser({
-        username,
-        name: resolvedDisplayName,
-        email: email ?? null,
-        emailVerified: emailVerified ?? false,
-        password: null,
-      });
-    } else if (emailVerified && !user.emailVerified) {
-      user = await this.usersService.updateUser(user.id, {
-        emailVerified: true,
-      });
-    }
-
-    if (!user) {
-      throw new InternalServerErrorException('사용자 생성에 실패했습니다.');
-    }
-
+    // 3. 계정 연결
     await this.usersService.linkOAuthAccount(user.id, provider, providerId);
 
     return user;
-  }
-
-  private generateUsername(
-    provider: OAuthProvider,
-    providerId: string,
-  ): string {
-    const prefix = provider.toString().toLowerCase();
-    const suffix = providerId.slice(-8);
-    return `${prefix}_user${suffix}`;
   }
 }
